+++ demo/nextjs/pages/_app.tsx
@@ -1,5 +1,6 @@
 import '@react-pdf-viewer/core/lib/styles/index.css';
 import '@react-pdf-viewer/default-layout/lib/styles/index.css';
+import '@react-pdf-viewer/highlight/lib/styles/index.css';
 
 import '../styles/index.css';